Refactor the command update function to the commands module
