+++ lib/commandHandlers.js
@@ -26,7 +26,12 @@
 var async = require('async'),
     logger = require('logops'),
     iotAgentLib = require('iotagent-node-lib'),
-    mqttClient;
+    constants = require('./constants'),
+    mqttClient,
+    config,
+    context = {
+        op: 'IoTAgentMQTT.Commands'
+    };
 
 /**
  * Generate a function that executes the given command in the device.
@@ -81,9 +86,48 @@ function commandHandler(id, type, service, subservice, attributes, callback) {
     });
 }
 
-function init(newMqttClient) {
-    mqttClient = newMqttClient;
+/**
+ * Process an update in the state of a command with information coming from the device.
+ *
+ * @param {String} apiKey           API Key corresponding to the Devices configuration.
+ * @param {String} deviceId         Id of the device to be updated.
+ * @param {Object} device           Device object containing all the information about a device.
+ * @param {Object} messageObj       JSON object sent using MQTT.
+ */
+function updateCommand(apiKey, deviceId, device, messageObj) {
+    var commandList = Object.keys(messageObj),
+        commandUpdates = [];
+
+    for (var i = 0; i < commandList.length; i++) {
+        commandUpdates.push(
+            async.apply(iotAgentLib.setCommandResult,
+                device.name,
+                config.iota.defaultResource,
+                apiKey,
+                commandList[i],
+                messageObj[commandList[i]],
+                constants.COMMAND_STATUS_COMPLETED,
+                device)
+        );
+    }
+
+    async.series(commandUpdates, function(error) {
+        if (error) {
+            logger.error(context,
+                'Couldn\'t update command status in the Context broker for device [%s] with apiKey [%s]: %s',
+                device.id, apiKey, error);
+        } else {
+            logger.debug(context, 'Single measure for device [%s] with apiKey [%s] successfully updated',
+                device.id, apiKey);
+        }
+    });
 }
 
+function init(newMqttClient, newConfig) {
+    mqttClient = newMqttClient;
+    config = newConfig;
+}
+
+exports.updateCommand = updateCommand;
 exports.handler = commandHandler;
 exports.init = init;
