+++ lib/iotagent-mqtt.js
@@ -274,43 +274,6 @@ function parseMessage(message) {
 }
 
 /**
- * Process an update in the state of a command with information coming from the device.
- *
- * @param {String} apiKey           API Key corresponding to the Devices configuration.
- * @param {String} deviceId         Id of the device to be updated.
- * @param {Object} device           Device object containing all the information about a device.
- * @param {Object} messageObj       JSON object sent using MQTT.
- */
-function updateCommand(apiKey, deviceId, device, messageObj) {
-    var commandList = Object.keys(messageObj),
-        commandUpdates = [];
-
-    for (var i = 0; i < commandList.length; i++) {
-        commandUpdates.push(
-            async.apply(iotAgentLib.setCommandResult,
-                device.name,
-                config.iota.defaultResource,
-                apiKey,
-                commandList[i],
-                messageObj[commandList[i]],
-                constants.COMMAND_STATUS_COMPLETED,
-                device)
-        );
-    }
-
-    async.series(commandUpdates, function(error) {
-        if (error) {
-            logger.error(context,
-                'Couldn\'t update command status in the Context broker for device [%s] with apiKey [%s]: %s',
-                device.id, apiKey, error);
-        } else {
-            logger.debug(context, 'Single measure for device [%s] with apiKey [%s] successfully updated',
-                device.id, apiKey);
-        }
-    });
-}
-
-/**
  * Handles an incoming MQTT message, extracting the API Key, device Id and attribute to update (in the case of single
  * measures) from the MQTT topic.
  *
@@ -332,7 +295,7 @@ function mqttMessageHandler(topic, message) {
             } else if (topicInformation[4]) {
                 singleMeasure(apiKey, deviceId, topicInformation[4], device, message);
             } else if (topicInformation[3] === constants.CONFIGURATION_COMMAND_UPDATE) {
-                updateCommand(apiKey, deviceId, device, parsedMessage);
+                commandHandler.updateCommand(apiKey, deviceId, device, parsedMessage);
             } else if (parsedMessage && typeof parsedMessage === 'object') {
                 multipleMeasures(apiKey, deviceId, device, parsedMessage);
             } else {
@@ -488,7 +451,7 @@ function start(newConfig, callback) {
             mqttClient.on('connect', function() {
                 logger.info(context, 'MQTT Client connected');
                 recreateSubscriptions(callback);
-                commandHandler.init(mqttClient);
+                commandHandler.init(mqttClient, config);
             });
         }
     });
